requirements.txt
torch
tensorflow
numpy
transformers
pillow
matplotlib